refactor(phase-3a): extract JiraProjectService
First slice of the jira_client.py god-class decomposition (ADR-002).
Five project-domain methods move from ``jira_client.py`` (2,852 LOC)
into a focused ``JiraProjectService`` exposed as ``self.projects``.
``JiraClient`` keeps a thin delegator for each method so existing
callers work unchanged.

Methods moved
-------------
- get_projects
- get_issue_types
- get_project_roles
- get_project_permission_scheme
- get_project_metadata_enhanced

Numbers
-------
- jira_client.py: 2852 -> 2610 LOC
- jira_project_service.py: 0 -> 316 LOC

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean
- ruff check / ruff format: clean
