+++ src/clients/jira_client.py
@@ -28,7 +28,6 @@ from src.utils.config_validation import ConfigurationValidationError, SecurityVa
 from src.utils.performance_optimizer import (
     PerformanceOptimizer,
     StreamingPaginator,
-    cached,
     rate_limited,
 )
 from src.utils.rate_limiter import create_jira_rate_limiter
@@ -240,6 +239,12 @@ class JiraClient:
         self.batch_size = batch_size
         self.parallel_workers = max_workers
 
+        # Service composition (Phase 3a of ADR-002 — see ADR for the
+        # decomposition plan).
+        from src.clients.jira_project_service import JiraProjectService
+
+        self.projects = JiraProjectService(self)
+
         # Connect to Jira
         self._connect()
         self._patch_jira_client()
@@ -319,128 +324,12 @@ class JiraClient:
         raise JiraAuthenticationError(msg) from None
 
     def get_projects(self) -> list[dict[str, Any]]:
-        """Get all projects from Jira with enriched metadata."""
-        if not self.jira:
-            msg = "Jira client is not initialized"
-            raise JiraConnectionError(msg)
-
-        def _sanitize_str(value: Any) -> str:
-            if value is None:
-                return ""
-            return str(value)
-
-        projects_details: list[dict[str, Any]] = []
-
-        try:
-            projects = self.jira.projects()
-            if not projects:
-                logger.warning("No projects found in Jira")
-                return []
-
-            for project in projects:
-                detail = None
-                try:
-                    detail = self.jira.project(project.id)
-                except Exception as exc:
-                    logger.debug(
-                        "Failed to fetch detailed project metadata for %s: %s",
-                        project.key,
-                        exc,
-                    )
-
-                raw_detail = getattr(detail, "raw", {}) or {}
-
-                project_category = raw_detail.get("projectCategory") or {}
-                if not isinstance(project_category, dict):
-                    project_category = {}
-
-                category_name = _sanitize_str(project_category.get("name")) if project_category else ""
-                category_id = _sanitize_str(project_category.get("id")) if project_category else ""
-
-                lead_info = raw_detail.get("lead") or {}
-                if not isinstance(lead_info, dict):
-                    lead_info = {}
-
-                lead_login = lead_info.get("name") or lead_info.get("key")
-                lead_display = lead_info.get("displayName")
-
-                avatar_urls = raw_detail.get("avatarUrls") or {}
-                if not isinstance(avatar_urls, dict):
-                    avatar_urls = {}
-                preferred_avatar_url = ""
-                for size_key in ("128x128", "64x64", "48x48", "32x32", "24x24", "16x16"):
-                    candidate = avatar_urls.get(size_key)
-                    if candidate:
-                        preferred_avatar_url = str(candidate)
-                        break
-
-                project_type = raw_detail.get("projectTypeKey") or getattr(project, "projectTypeKey", None)
-
-                browse_url = f"{self.base_url}/browse/{project.key}"
-
-                description = raw_detail.get("description") or ""
-                if description is None:
-                    description = ""
-
-                projects_details.append(
-                    {
-                        "key": project.key,
-                        "name": project.name,
-                        "id": project.id,
-                        "project_type_key": project_type,
-                        "project_category": project_category,
-                        "project_category_name": category_name,
-                        "project_category_id": category_id,
-                        "description": description,
-                        "lead": lead_login,
-                        "lead_display": lead_display,
-                        "avatar_urls": avatar_urls,
-                        "avatar_url": preferred_avatar_url,
-                        "url": raw_detail.get("self"),
-                        "browse_url": browse_url,
-                        "archived": raw_detail.get("archived", False),
-                    },
-                )
-
-            return projects_details
-        except Exception as e:
-            error_msg = f"Failed to get projects: {e!s}"
-            logger.exception(error_msg)
-            raise JiraApiError(error_msg) from e
+        """Thin delegator over ``self.projects.get_projects``."""
+        return self.projects.get_projects()
 
     def get_issue_types(self) -> list[dict[str, Any]]:
-        """Get all issue types from Jira.
-
-        Returns:
-            List of issue type dictionaries with id, name, and description
-
-        Raises:
-            JiraApiError: If the API request fails
-
-        """
-        if not self.jira:
-            msg = "Jira client is not initialized"
-            raise JiraConnectionError(msg)
-
-        try:
-            issue_types = self.jira.issue_types()
-            result = [
-                {
-                    "id": issue_type.id,
-                    "name": issue_type.name,
-                    "description": issue_type.description,
-                }
-                for issue_type in issue_types
-            ]
-
-            if not issue_types:
-                logger.warning("No issue types found in Jira")
-
-            return result
-        except Exception as e:
-            error_msg = f"Failed to get issue types: {e!s}"
-            logger.exception(error_msg)
-            raise JiraApiError(error_msg) from e
+        """Thin delegator over ``self.projects.get_issue_types``."""
+        return self.projects.get_issue_types()
 
     def get_all_issues_for_project(
         self,
@@ -854,99 +743,12 @@ class JiraClient:
             raise JiraApiError(error_msg) from e
 
     def get_project_roles(self, project_key: str) -> list[dict[str, Any]]:
-        """Retrieve Jira project roles and their actors for a project."""
-        if not project_key:
-            return []
-
-        logger.debug("Fetching Jira project roles for '%s'", project_key)
-
-        try:
-            role_map_response = self._make_request(
-                f"/rest/api/2/project/{project_key}/role",
-            )
-            if role_map_response.status_code != HTTP_OK:
-                msg = f"Failed to fetch Jira project roles for {project_key}: HTTP {role_map_response.status_code}"
-                raise JiraApiError(msg)
-
-            role_map = role_map_response.json() or {}
-            roles: list[dict[str, Any]] = []
-
-            for role_name, role_url in role_map.items():
-                if not isinstance(role_url, str):
-                    continue
-
-                detail_path = role_url
-                if role_url.startswith(self.base_url):
-                    detail_path = role_url[len(self.base_url) :]
-                if not detail_path.startswith("/"):
-                    detail_path = f"/{detail_path}"
-
-                detail_response = self._make_request(detail_path)
-                if detail_response.status_code != HTTP_OK:
-                    logger.warning(
-                        "Skipping Jira role '%s' for project '%s' due to HTTP %s",
-                        role_name,
-                        project_key,
-                        detail_response.status_code,
-                    )
-                    continue
-
-                detail = detail_response.json() or {}
-                actors = []
-                for actor in detail.get("actors", []):
-                    actors.append(
-                        {
-                            "type": actor.get("type"),
-                            "name": actor.get("name"),
-                            "displayName": actor.get("displayName"),
-                            "accountId": ((actor.get("actorUser") or {}).get("accountId") or actor.get("accountId")),
-                            "userKey": ((actor.get("actorUser") or {}).get("key") or actor.get("userKey")),
-                            "groupName": ((actor.get("actorGroup") or {}).get("name") or actor.get("groupName")),
-                        },
-                    )
-
-                roles.append(
-                    {
-                        "id": detail.get("id"),
-                        "name": detail.get("name") or role_name,
-                        "description": detail.get("description"),
-                        "actors": actors,
-                    },
-                )
-
-            logger.debug(
-                "Discovered %s Jira project roles for '%s'",
-                len(roles),
-                project_key,
-            )
-            return roles
-        except JiraCaptchaError, JiraAuthenticationError, JiraConnectionError:
-            raise
-        except Exception as e:
-            error_msg = f"Failed to fetch Jira project roles for {project_key}: {e!s}"
-            logger.exception(error_msg)
-            raise JiraApiError(error_msg) from e
+        """Thin delegator over ``self.projects.get_project_roles``."""
+        return self.projects.get_project_roles(project_key)
 
     def get_project_permission_scheme(self, project_key: str) -> dict[str, Any]:
-        """Return the permission scheme applied to a Jira project."""
-        if not self.jira:
-            msg = "Jira client is not initialized"
-            raise JiraConnectionError(msg)
-
-        path = f"{self.base_url}/rest/api/2/project/{project_key}/permissionscheme"
-        logger.debug("Fetching Jira permission scheme for project '%s'", project_key)
-
-        try:
-            response = self.jira._session.get(path)
-            if response.status_code == HTTP_NOT_FOUND:
-                return {}
-            response.raise_for_status()
-            payload = response.json()
-            return payload if isinstance(payload, dict) else {}
-        except Exception as exc:
-            error_msg = f"Failed to fetch permission scheme for {project_key}: {exc!s}"
-            logger.exception(error_msg)
-            raise JiraApiError(error_msg) from exc
+        """Thin delegator over ``self.projects.get_project_permission_scheme``."""
+        return self.projects.get_project_permission_scheme(project_key)
 
     def get_issue_count(self, project_key: str) -> int:
         """Get the total number of issues in a project.
@@ -2803,50 +2605,6 @@ class JiraClient:
 
         return {key: user_dict[key] for key in user_keys if key in user_dict}
 
-    @cached(ttl=3600)  # Cache for 1 hour
     def get_project_metadata_enhanced(self, project_key: str) -> dict[str, Any]:
-        """Get comprehensive project metadata with caching."""
-        try:
-            project = self.jira.project(project_key)
-
-            # Get additional metadata
-            issue_types = self.jira.createmeta_issuetypes(project.key)
-            statuses = self.jira.project_status(project.key)
-
-            return {
-                "project": {
-                    "id": project.id,
-                    "key": project.key,
-                    "name": project.name,
-                    "description": getattr(project, "description", ""),
-                    "lead": (
-                        getattr(project, "lead", {}).get("name", "Unknown") if hasattr(project, "lead") else "Unknown"
-                    ),
-                    "project_type_key": getattr(project, "projectTypeKey", "software"),
-                },
-                "issue_types": [
-                    {
-                        "id": it.id,
-                        "name": it.name,
-                        "description": getattr(it, "description", ""),
-                        "subtask": getattr(it, "subtask", False),
-                    }
-                    for it in issue_types
-                ],
-                "statuses": [
-                    {
-                        "id": status.id,
-                        "name": status.name,
-                        "description": getattr(status, "description", ""),
-                        "category": getattr(status, "statusCategory", {}).get(
-                            "name",
-                            "Unknown",
-                        ),
-                    }
-                    for status in statuses
-                ],
-            }
-        except Exception as e:
-            error_msg = f"Failed to get enhanced project metadata for {project_key}: {e}"
-            logger.exception(error_msg)
-            raise JiraApiError(error_msg) from e
+        """Thin delegator over ``self.projects.get_project_metadata_enhanced``."""
+        return self.projects.get_project_metadata_enhanced(project_key)
